% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_vdj_functions.R
\name{cellranger_vdj_filter}
\alias{cellranger_vdj_filter}
\title{Filter Cell Ranger VDJ}
\usage{
cellranger_vdj_filter(
  contig_annotations,
  filterTrue = c("full_length", "high_confidence"),
  min.umis = 0
)
}
\arguments{
\item{contig_annotations}{Dataframe containing contig_annotations}

\item{filterTrue}{Vector of which "Boolean columns" in cellranger contig_annotation should be true ("full_length", "high_confidence", "productive")}

\item{min.umis}{Integer threshold for minimum UMI count required to be kept}
}
\value{
data.frame containing contig annotations
}
\description{
Filter by boolean columns
}
