% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_vdj_functions.R
\name{cellranger_vdj_load}
\alias{cellranger_vdj_load}
\title{Load Cell Ranger VDJ}
\usage{
cellranger_vdj_load(
  path,
  consensus = TRUE,
  annotation_file = "outs/filtered_contig_annotations.csv",
  consensus_file = "outs/consensus_annotations.csv",
  ...
)
}
\arguments{
\item{path}{Path to Cell Ranger output (root folder, not "outs" folder)}

\item{consensus}{Boolean if contigs assigned to consensus sequences should use the consensus annotation}

\item{annotation_file}{contig_annotation file to use ("all" versus "filtered")}

\item{consensus_file}{consensus_annotation file to use}

\item{...}{Passes variables on to \code{\link{consensus_merge}}}
}
\value{
data.frame containing contig annotations
}
\description{
Load Cell Ranger VDJ
}
