% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_vdj_functions.R
\name{cellranger_vdj_order}
\alias{cellranger_vdj_order}
\title{Order Cell Ranger VDJ calls}
\usage{
cellranger_vdj_order(contig_annotations)
}
\arguments{
\item{contig_annotations}{Dataframe containing contig_annotations}
}
\value{
data.frame containing contig annotations
}
\description{
To determine which contigs are most likely to be "primary" (~expressed) loci
}
