% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_vdj_functions.R
\name{cellranger_vdj_parse_by_cell}
\alias{cellranger_vdj_parse_by_cell}
\title{Parse cellranger contigs}
\usage{
cellranger_vdj_parse_by_cell(contig_annotations)
}
\arguments{
\item{contig_annotations}{Dataframe containing contig_annotations}
}
\value{
data.frame containing contig annotations
}
\description{
Parse cellranger contigs contig_annotation table to call VDJ on a per
cell-barcode basis
}
