% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotype_functions.R
\name{clonotype_filter}
\alias{clonotype_filter}
\title{Filter clonotypes by number of cells}
\usage{
clonotype_filter(clonotypes, cells_min = 5, nonclonal_name = "Polyclonal")
}
\arguments{
\item{clonotypes}{Vector of clonotype assignment to each cell}

\item{cells_min}{Integer of number of cells required to be considered a clone}

\item{nonclonal_name}{Character string for labelling clonotypes that do not fulfill filtering criteria}
}
\value{
Returns a vector of same length as input clonotypes with filtered clonotype assignments
}
\description{
Filter clonotypes by number of cells
}
