% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotype_functions.R
\name{clonotype_top}
\alias{clonotype_top}
\title{Filter top clonotypes by rank}
\usage{
clonotype_top(
  clonotypes,
  clonotypes_max = 10,
  cells_min = 1,
  nonclonal_name = "Polyclonal",
  ...
)
}
\arguments{
\item{clonotypes}{Vector of clonotype assignment to each cell}

\item{clonotypes_max}{Max number of clonotypes to return}

\item{cells_min}{Integer of number of cells required to be considered a clone}

\item{...}{Passed on to clonotype_filter}
}
\description{
Filter top clonotypes by rank
}
