% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_merge.R
\name{consensus_merge}
\alias{consensus_merge}
\title{Merge consensus VDJ annotation into contigs}
\usage{
consensus_merge(
  data_annotation,
  data_consensus,
  mutual_columns = c("length", "reads", "umis"),
  consensus_suffix = ".consensus"
)
}
\arguments{
\item{data_annotation}{Contig annotations from Cell Ranger output}

\item{data_consensus}{Consensus contigs from Cell Ranger output}

\item{mutual_columns}{Columns that are included from both outputs (set by mutual_columns)}

\item{consensus_suffix}{Suffix given to mutual_columns from the consensus data}
}
\value{
data.frame containing contig annotations
}
\description{
Merges consensus VDJ annotations into contig annotation data.frame to allow incomplete
to be called as most likely clonotype
}
