% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{paste_vdj}
\alias{paste_vdj}
\title{Make VDJ string}
\usage{
paste_vdj(x, sep = ":")
}
\arguments{
\item{x}{vector of values}

\item{sep}{seperator in concatenated string}
}
\value{
Concatenated string where empty values have been removed
}
\description{
Concatenate values of multiple columns into a single string after removing empty values by \code{no_empty}
}
