% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vdj_circos.R
\name{plot_vdj_circos}
\alias{plot_vdj_circos}
\title{Circos plot of V and J segments}
\usage{
plot_vdj_circos(
  Vsegment,
  Jsegment,
  Dsegment = c(),
  clone = c(),
  polyclonal_label = "Polyclonal",
  polyclonal_color = "#00000020",
  start_degree = 270,
  clone_opacity = 80,
  clone_colors,
  link_lwd = 0.25,
  label_threshold = 2
)
}
\arguments{
\item{Vsegment}{Vector of V segment assignment}

\item{Jsegment}{Vector of J segment assignment}

\item{Dsegment}{Vector of D segment assignment}

\item{clone}{Vector of clone assignment}

\item{polyclonal_label}{What label is used to tag polyclonal (within clone vector)}

\item{polyclonal_color}{Color for polyclonal connections}

\item{start_degree}{Where on the circle should the first segment start}

\item{clone_opacity}{Opacity of clonal connections}

\item{clone_colors}{Named vector of colors (names corresponds to values in clone vector)}

\item{link_lwd}{Line width of connections}
}
\description{
Circos plot of V and J segments
}
