% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vdj_circos.R
\name{plot_vdj_circos_seurat}
\alias{plot_vdj_circos_seurat}
\title{Circos plot of V and J segments from Seurat object}
\usage{
plot_vdj_circos_seurat(
  object,
  Vsegment,
  Jsegment,
  Dsegment = NULL,
  clone,
  cells = c(),
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{Vsegment}{Vector of V segment assignment}

\item{Jsegment}{Vector of J segment assignment}

\item{Dsegment}{Vector of D segment assignment (not currently implemented)}

\item{clone}{Vector of clone assignment}

\item{cells}{Which cells should be included}
}
\description{
Circos plot of V and J segments from Seurat object
}
