% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vdj_clones_dimred.R
\name{plot_vdj_clones_dimred}
\alias{plot_vdj_clones_dimred}
\title{Plot VDJ clones on DimPlot}
\usage{
plot_vdj_clones_dimred(
  x,
  y,
  clone,
  clone_colors = c(),
  red.dims = c("dim1", "dim2"),
  polyclonal_label = "Polyclonal",
  polyclonal_color = "#6666FF",
  clone_size = 2,
  polyclonal_size = 0.75
)
}
\value{
ggplot object
}
\description{
Plot VDJ clones on DimPlot
}
