% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vdj_clones_dimred.R
\name{plot_vdj_clones_dimred_seurat}
\alias{plot_vdj_clones_dimred_seurat}
\title{Plot VDJ clones on DimPlot seurat}
\usage{
plot_vdj_clones_dimred_seurat(object, clone, reduction = "umap", ...)
}
\arguments{
\item{object}{Seurat object}
}
\value{
ggplot object
}
\description{
Plot VDJ clones on DimPlot seurat
}
