% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{no_empty}
\alias{no_empty}
\title{Determines if a field from cellranger vdj output is empty}
\usage{
no_empty(x, empty.values = c("", "None"))
}
\arguments{
\item{x}{Input vector}

\item{empty.values}{Vector of values considered "empty"}
}
\value{
Vector where empty values have been removed
}
\description{
Considered empty if equal to "" or "None"
}
